backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
lightkurve==2.4.2
astroquery==0.4.7
scikit-learn==1.3.2
scipy==1.11.4
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
astropy==5.3.4
joblib==1.3.2
cachetools==5.3.2
websockets==11.0.3


# Dependencias para el script de subida a S3
boto3>=1.28.0
python-dotenv>=1.0.0
botocore>=1.31.0
pydantic>=2.10.7
pydantic-settings>=2.10.7